@@ -1,27 +1,14 @@
 import CssBaseline from "@mui/material/CssBaseline";
-import Head from "next/head";
 import React from "react";
-import { GlobalFooter, GlobalNav, Seo } from "../../pattern";
+import { GlobalFooter, GlobalNav } from "../../pattern";
 
 export default function DefaultTemplate({
   children,
-  title,
-  metaDesc,
 }: {
   children?: React.ReactNode;
-  title?: string;
-  metaDesc?: string;
 }) {
-  const titleTransform = title && `${title} - `,
-    pageTitle = `${titleTransform}Millennial Realty & Investments LLC`;
-
   return (
     <>
-      <Head>
-        <title>{pageTitle}</title>
-        <meta name="viewport" content="initial-scale=1, width=device-width" />
-        <Seo title={pageTitle} description={metaDesc} />
-      </Head>
       <CssBaseline />
       <GlobalNav />
       {children}
@@ -29,8 +16,3 @@ export default function DefaultTemplate({
     </>
   );
 }
-
-DefaultTemplate.defaultProps = {
-  title: "",
-  metaDesc: "",
-};
